remove python 2 support and other minor cleanup
