+++ rlp/sedes/raw.py
@@ -3,11 +3,10 @@ A sedes that does nothing. Thus, everything that can be directly encoded by RLP
 is serializable. This sedes can be used as a placeholder when deserializing
 larger structures.
 """
-
-
 from collections import Sequence
-from ..exceptions import SerializationError
-from ..utils import Atomic
+
+from rlp.exceptions import SerializationError
+from rlp.utils import Atomic
 
 
 def serializable(obj):
